# Inline attributes in joins to speed up performance in `/content` endpoint