--- a/app/models/dimensions/edition.rb
+++ b/app/models/dimensions/edition.rb
@@ -23,7 +23,6 @@ class Dimensions::Edition < ApplicationRecord
     latest_by_content_id(content_id, locale)
       .where.not(base_path: exclude_paths)
   end
-  scope :relevant_content, -> { where.not(document_type: DocumentType::IGNORED_TYPES) }
 
   def self.search(query)
     sql = <<~SQL