--- a/app/models/dimensions/edition.rb
+++ b/app/models/dimensions/edition.rb
@@ -24,15 +24,6 @@ class Dimensions::Edition < ApplicationRecord
       .where.not(base_path: exclude_paths)
   end
 
-  def self.search(query)
-    sql = <<~SQL
-      to_tsvector('english',title) @@ plainto_tsquery('english', :search_term) or
-      to_tsvector('english'::regconfig, replace((base_path)::text, '/'::text, ' '::text)) @@ plainto_tsquery('english', :search_term_without_slash)
-    SQL
-
-    where(sql, search_term: query, search_term_without_slash: query.tr('/', ' '))
-  end
-
   def promote!(old_edition)
     if old_edition
       old_edition.deprecate!