--- a/db/migrate/20190213115655_update_aggregations_search_last_thirty_days_to_version_5.rb
+++ b/db/migrate/20190213115655_update_aggregations_search_last_thirty_days_to_version_5.rb
@@ -1,0 +1,5 @@
+class UpdateAggregationsSearchLastThirtyDaysToVersion5 < ActiveRecord::Migration[5.2]
+  def change
+    update_view :aggregations_search_last_thirty_days, version: 5, revert_to_version: 4, materialized: true
+  end
+end