--- a/db/migrate/20190213141732_update_aggregations_search_last_months_to_version_3.rb
+++ b/db/migrate/20190213141732_update_aggregations_search_last_months_to_version_3.rb
@@ -1,0 +1,5 @@
+class UpdateAggregationsSearchLastMonthsToVersion3 < ActiveRecord::Migration[5.2]
+  def change
+    update_view :aggregations_search_last_months, version: 3, revert_to_version: 2, materialized: true
+  end
+end