--- a/db/migrate/20190213155920_add_indexes_view_aggregations_search_last_three_months.rb
+++ b/db/migrate/20190213155920_add_indexes_view_aggregations_search_last_three_months.rb
@@ -4,8 +4,8 @@ class AddIndexesViewAggregationsSearchLastThreeMonths < ActiveRecord::Migration[
     execute "create index aggregations_search_last_three_months_gin_title on aggregations_search_last_three_months  using gin(to_tsvector('english', title ))"
 
     add_index :aggregations_search_last_three_months, :upviews, name: :search_last_three_months_gin_base_path_upviews
-    add_index :aggregations_search_last_three_months, [:organisation_id, :upviews], name: :search_last_three_months_gin_base_path_organisation_id
-    add_index :aggregations_search_last_three_months, [:document_type, :upviews], name: :search_last_three_months_gin_base_path_document_type
+    add_index :aggregations_search_last_three_months, %i[organisation_id upviews], name: :search_last_three_months_gin_base_path_organisation_id
+    add_index :aggregations_search_last_three_months, %i[document_type upviews], name: :search_last_three_months_gin_base_path_document_type
   end
 
   def down