… db/schema.rb
… db/views/aggregations_search_last_thirty_days_v05.sql
… db/views/aggregations_search_last_three_months_v05.sql
… db/views/aggregations_search_last_twelve_months_v05.sql
… lib/tasks/repopulate_pdf_count.rake
… spec/models/dimensions/edition_spec.rb